add metrics collection post hook
This commit adds a script that collects metrics from the build output.

The metrics have to be printed out as a json object in a single line,
e.g.,

{ "result" : foo }

The format doesn't matter, as long as it can be set as value foo
into { "metrics" : { "<app>" : { "<board>" : foo }}}.

Any line starting with "{" will be considered. If it doesn't json parse
correctly (json.loads() throws an exception), the line is ignored.

All metrics will be written to "metrics.json", looking like this:

    {
        "metrics": {
            "tests/bench_msg_pingpong": {
                "native:gnu": {
                    "result": 297007
                },
                "native:llvm": {
                    "result": 171431
                }
            },
            "tests/bench_mutex_pingpong": {
                "native:gnu": {
                    "result": 92752
                },
                "native:llvm": {
                    "result": 336211
                }
            },
        ...
        }
    }